third_party/WebKit/LayoutTests/http/tests/csspaint/registered-properties-in-custom-paint-expected.txt

CONSOLE MESSAGE: line 14: --length: [CSSStyleValue=10px]
CONSOLE MESSAGE: line 14: --length-initial: [CSSStyleValue=20px]
CONSOLE MESSAGE: line 14: --number: [CSSStyleValue=10]
This tests the style information in the paint callback.

See the devtools console for test output. The console should log:

--length: [CSSStyleValue=10px]
--length-initial: [CSSStyleValue=20px]
--number: [CSSStyleValue=10]
